Remove `include/grpcpp/impl/codegen/string_ref.h`
