+++ test/cpp/ios/CronetTests/TestHelper.h
@@ -25,9 +25,9 @@
 #import <sstream>
 
 #import <grpc/support/time.h>
-#import <grpcpp/impl/codegen/string_ref.h>
 #import <grpcpp/support/client_interceptor.h>
 #import <grpcpp/support/config.h>
+#import <grpcpp/support/string_ref.h>
 
 #import "src/proto/grpc/testing/echo.grpc.pb.h"
